minor changes
git-svn-id: file:///home/roberto/Git/svn@7630 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 656b8822e84b7215837e1476d5c26dba9b7f3670

+++ trunk/xmipp/libraries/data/multidim_array.cpp
@@ -36,7 +36,7 @@ std::ostream& operator<<(std::ostream& ostrm,
     else
         ostrm << std::endl;
 
-    for (int l = 0; l < NSIZE(v); l++)
+    for (size_t l = 0; l < NSIZE(v); l++)
     {
         if (NSIZE(v)>1)
             ostrm << "Image No. " << l << std::endl;
